Add books controller and serializer for response
