+++ app/controllers/users/current_user_controller.rb
@@ -1,0 +1,11 @@
+module Users
+  class CurrentUserController < ApplicationController
+    before_action :authenticate_user!
+
+    def index
+      render json: {
+               data: UserSerializer.new(current_user).serializable_hash[:data][:attributes], status: :ok,
+             }
+    end
+  end
+end
