+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
-  get "/current_user", to: "current_user#index"
+  get "/current_user", to: "users/current_user#index"
   devise_for :users, path: "", path_names: {
                        sign_in: "login",
                        sign_out: "logout",
@@ -9,5 +9,6 @@ Rails.application.routes.draw do
                        sessions: "users/sessions",
                        registrations: "users/registrations",
                      }
-  resources :users, only: %i[index]
+  resources :users, only: %i[index], controller: "users/users"
+  resources :books
 end
